json bindings. float64s.

+++ go/models/gear.go
@@ -5,10 +5,10 @@ type Gear struct {
 	HP      int
 	Attack  int
 	Defense int
-	Speed   int
-	Crit    int
-	Dodge   int
-	Block   int
+	Speed   float64
+	Crit    float64
+	Dodge   float64
+	Block   float64
 	SlotId  int
 	Rarity  int
 }
